+++ flash-core/src/main/java/cn/enilu/flash/dao/sales/Sla20Repository.java
@@ -8,7 +8,7 @@ import cn.enilu.flash.dao.BaseRepository;
 
 public interface Sla20Repository  extends BaseRepository<Sla20,Long>{
 	
-	public Sla20 findBySla20002AndSla20006(int projectUID, int customerUID);
+	public Sla20 findBySla20002AndSla20006(Long projectUID, Long customerUID);
 	
 	
 }
